feat: external access page for mobile devices
