+++ src/meetings/views/MeetingMainView.tsx
@@ -4,36 +4,22 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import React, { lazy, ReactElement, Suspense, useEffect } from 'react';
+import React, { ReactElement, Suspense, useEffect } from 'react';
 
 import { createMemoryHistory } from 'history';
 import { Route, Router, Switch } from 'react-router-dom';
 
+import {
+	LazyAccessPageView,
+	LazyExternalAccessPage,
+	LazyInfoPage,
+	LazyMeetingAccessPageView,
+	LazyMeetingSkeleton
+} from './LazyLoadedPages';
 import ShimmerEntryMeetingView from './shimmers/ShimmerEntryMeetingView';
 import { MEETINGS_ROUTES, ROUTES } from '../../hooks/useRouting';
 import { MeetingsApi } from '../../network';
 import useStore from '../../store/Store';
-
-const LazyAccessPageView = lazy(
-	() => import(/* webpackChunkName: "MeetingAccessPageView" */ './AccessPage')
-);
-
-const LazyMeetingSkeleton = lazy(
-	() => import(/* webpackChunkName: "MeetingSkeleton" */ './MeetingSkeleton')
-);
-
-const LazyInfoPage = lazy(() => import(/* webpackChunkName: "InfoPage" */ './InfoPage'));
-
-const LazyMeetingExternalAccessPage = lazy(
-	() =>
-		import(
-			/* webpackChunkName: "MeetingExternalAccessPage" */ '../components/meetingAccessPoint/MeetingExternalAccessPage'
-		)
-);
-
-const LazyMeetingAccessPageView = lazy(
-	() => import(/* webpackChunkName: "MeetingAccessPageView" */ './MeetingAccessPageView')
-);
 
 const AccessPageView = (): ReactElement => (
 	<Suspense fallback={<ShimmerEntryMeetingView />}>
@@ -55,7 +41,7 @@ const InfoPage = (): ReactElement => (
 
 const MeetingExternalAccessPage = (): ReactElement => (
 	<Suspense fallback={<ShimmerEntryMeetingView />}>
-		<LazyMeetingExternalAccessPage />
+		<LazyExternalAccessPage />
 	</Suspense>
 );
